+++ app/actions/fetchReservation.ts
@@ -11,9 +11,12 @@ export async function fetchReservations() {
         reservationNumber: true,
       },
     });
-    return reservations;
+    return reservations.map((reservation) => ({
+      ...reservation,
+      date: new Date(reservation.date).toISOString(), // Ensure consistent date formatting
+    }));
   } catch (error) {
-    console.error("Erro ao buscar reservas:", error);
-    throw new Error("Falha ao buscar reservas");
+    console.error("Error fetching reservations:", error);
+    throw new Error("Failed to fetch reservations");
   }
 }
