+++ app/adm/member-manage/page.tsx
@@ -1,5 +1,5 @@
-import AdmNavbar from "@/components/AdmNavBar";
-import MemberManagementTable from "@/components/MemberTable";
+import AdmNavbar from "@/components/Navbar/AdmNavBar";
+import MemberManagementTable from "@/components/Tables/MemberTable";
 
 
 export default function NewReserve() {
