+++ app/guide/bath/page.tsx
@@ -1,7 +1,7 @@
 import BathFacilities from "@/components/BathFacilities";
-import Footer from "@/components/Footer";
-import FooterM from "@/components/FooterM";
-import NavBar from "@/components/NavBar";
+import Footer from "@/components/Footer/Footer";
+import FooterM from "@/components/Footer/FooterM";
+import NavBar from "@/components/Navbar/NavBar";
 
 
 export default function BathRoom() {
