+++ app/how-to/page.tsx
@@ -1,7 +1,7 @@
-import Footer from "@/components/Footer";
-import FooterM from "@/components/FooterM";
+import Footer from "@/components/Footer/Footer";
+import FooterM from "@/components/Footer/FooterM";
 import HowToDo from "@/components/HowToDo";
-import NavBar from "@/components/NavBar";
+import NavBar from "@/components/Navbar/NavBar";
 
 
 export default function HowTo() {
