+++ app/page.tsx
@@ -1,8 +1,8 @@
 
 import Banner from "@/components/Banner";
-import Footer from "@/components/Footer";
-import FooterM from "@/components/FooterM";
-import NavBarW from "@/components/NavBarW";
+import Footer from "@/components/Footer/Footer";
+import FooterM from "@/components/Footer/FooterM";
+import NavBarW from "@/components/Navbar/NavBarW";
 
 export default function Home() {
   return (
