+++ app/register/submit2/page.tsx
@@ -1,8 +1,8 @@
 
 
-import Footer from "@/components/Footer";
-import FooterM from "@/components/FooterM";
-import NavBarW from "@/components/NavBarW";
+import Footer from "@/components/Footer/Footer";
+import FooterM from "@/components/Footer/FooterM";
+import NavBarW from "@/components/Navbar/NavBarW";
 import NewUserPage from "@/components/NewUserPage";
 
 export default function NewUser() {
