+++ app/reserve/steps/page.tsx
@@ -1,8 +1,8 @@
 
 
-import Footer from "@/components/Footer";
-import FooterM from "@/components/FooterM";
-import NavBar from "@/components/NavBar";
+import Footer from "@/components/Footer/Footer";
+import FooterM from "@/components/Footer/FooterM";
+import NavBar from "@/components/Navbar/NavBar";
 import Step1 from "@/components/Step1";
 import Step2 from "@/components/Step2";
 import Step3 from "@/components/Step3";
