+++ components/Navbar/NavBarW.tsx
@@ -3,21 +3,21 @@ import Image from 'next/image';
 import Link from 'next/link';
 import { useState } from 'react';
 
-const NavBar = () => {
+const NavBarW = () => {
   const [isOpen, setIsOpen] = useState(false);
 
   return (
-    <nav className=" z-10 w-full bg-transparent px-6 py-4">
+    <nav className="absolute z-10 w-full bg-transparent px-6 py-4">
       <div className="flex items-center justify-between">
         
         {/* Logo */}
-        <div className=" flex items-center">
+        <div className="  flex items-center">
           <Image
             src="/SaunaLogo.png"
             alt="Logo"
             width={60}
             height={65}
-            className="mr-4 invert"
+            className="mr-4 "
           />
         </div>
 
@@ -25,7 +25,7 @@ const NavBar = () => {
         <div className="relative flex lg:hidden">
           <button
             onClick={() => setIsOpen(!isOpen)}
-            className="text-gray-800 focus:outline-none"
+            className=" text-white focus:outline-none"
           >
             {/* Hamburger icon */}
             <svg
@@ -58,30 +58,30 @@ const NavBar = () => {
               <div className="my-4 border-t border-gray-300"></div>
 
               {/* Navigation Links in Specified Order */}
-              <Link href="/intro" className="block hover:text-gray-500">솔로사우나 레포 소개</Link>
+              <Link href="/intro" className="block  hover:text-gray-500">솔로사우나 레포 소개</Link>
               <Link href="/guide/women" className="block hover:text-gray-500">시설 안내</Link>
               <Link href="/how-to" className="block hover:text-gray-500">이용 방법</Link>
               <Link href="/comunity" className="block hover:text-gray-500">커뮤니티</Link>
               <Link href="/directions" className="block hover:text-gray-500">오시는 길</Link>
-              <Link href="#reservation" className="block hover:text-gray-500">예약하기</Link>
+              <Link href="/reserve" className="block hover:text-gray-500">예약하기</Link>
             </div>
           )}
         </div>
 
         {/* Full Navigation Links for Larger Screens */}
-        <div className="hidden space-x-8  text-gray-800 lg:flex lg:text-[15px] xl:text-[18px]">
+        <div className="hidden space-x-8 text-white  lg:flex lg:text-[15px] xl:text-[18px]">
           <Link href="/intro" className="hover:text-gray-500">솔로사우나 레포 소개</Link>
           <Link href="/guide/women" className="hover:text-gray-500">시설 안내</Link>
           <Link href="/how-to" className="hover:text-gray-500">이용 방법</Link>
           <Link href="/comunity" className="hover:text-gray-500">커뮤니티</Link>
           <Link href="/directions" className="hover:text-gray-500">오시는 길</Link>
-          <Link href="#reservation" className="hover:text-gray-500">예약하기</Link>
+          <Link href="/reserve" className="hover:text-gray-500">예약하기</Link>
         </div>
 
         {/* Authentication Links for Larger Screens */}
-        <div className=" hidden space-x-4 text-[18px] text-black lg:flex">
+        <div className=" hidden space-x-4 text-[18px] text-white lg:flex">
           <Link href="/login" className="hover:text-gray-500">로그인</Link>
-          <div className='border-l-2 border-l-black '></div>
+          <div className='border-l-2 border-l-white '></div>
           <Link href="/register/submit" className="hover:text-gray-500">회원 가입</Link>
         </div>
       </div>
@@ -89,4 +89,4 @@ const NavBar = () => {
   );
 };
 
-export default NavBar;
+export default NavBarW;
