+++ components/Tables/InquiresTable.tsx
